feat: add scan article
